Move OptionalQuery to axum_extra::query
