+++ axum-extra/src/extract/mod.rs
@@ -11,9 +11,6 @@ mod form;
 pub mod cookie;
 
 #[cfg(feature = "query")]
-mod optional_query;
-
-#[cfg(feature = "query")]
 mod query;
 
 #[cfg(feature = "multipart")]
@@ -34,10 +31,7 @@ pub use self::cookie::SignedCookieJar;
 pub use self::form::{Form, FormRejection};
 
 #[cfg(feature = "query")]
-pub use self::optional_query::{OptionalQuery, OptionalQueryRejection};
-
-#[cfg(feature = "query")]
-pub use self::query::{Query, QueryRejection};
+pub use self::query::{OptionalQuery, OptionalQueryRejection, Query, QueryRejection};
 
 #[cfg(feature = "multipart")]
 pub use self::multipart::Multipart;
